Split Splitfile generation into two modules
  - `common`: just generates a Splitfile from provenance data without
    replacing repos/images. Used by the parent repo to visualize Splitfiles
    without having to pull in pglast
  - `replacement`: current code that handles image injection

Factor code out in such a way that using `.common` doesn't pull in `Repository`
etc (passing in callbacks instead).
