+++ splitgraph/splitfile/generation/BUILD
@@ -1,0 +1,3 @@
+python_sources(
+    skip_black=True,
+)
